WIP Update fighting to use grid for proximity
Physics sphere gave massive lag, so we use the grid instead to have spatial
awareness of the entities on the map.

+++ Assets/06_Scripts/Entities/3_FightingData.cs
@@ -8,7 +8,7 @@ using UnityEngine;
 public class FightingData : ScriptableObject
 {
     [SerializeField]
-    private Ranges ranges = new Ranges(1.5f, 10f);
+    private Ranges ranges = new Ranges(1.5f, 10);
     public Ranges Ranges => ranges;
 
     [SerializeField]
@@ -24,9 +24,9 @@ public class FightingData : ScriptableObject
 public struct Ranges
 {
     public float attack;
-    public float vision;
+    public int vision;
 
-    public Ranges(float attack, float vision)
+    public Ranges(float attack, int vision)
     {
         this.attack = attack;
         this.vision = vision;
